+++ aad_encoder.c
@@ -370,7 +370,7 @@ static uint8_t AADEncodeProcessor_EncodeSample(
       break;
     default: AAD_ASSERT(0);
   }
-  idx = AAD_INNER_VAL(idx, 0, (int16_t)(AAD_NUM_TABLE_ELEMENTS(AAD_stepsize_table) - 1));
+  idx = AAD_INNER_VAL(idx, 0, (int32_t)(AAD_NUM_TABLE_ELEMENTS(AAD_stepsize_table) - 1));
 
   /* 計算結果の反映 */
   AAD_ASSERT(idx <= UINT8_MAX);
